moved viewholders inside adapter folder
