+++ app/src/main/java/co/develhope/meteoapp/today/adapter/TodayTitleViewHolder.kt
@@ -1,7 +1,8 @@
-package co.develhope.meteoapp.today
+package co.develhope.meteoapp.today.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import co.develhope.meteoapp.databinding.TitleTodayScreenBinding
+import co.develhope.meteoapp.today.TodayData
 
 class TodayTitleViewHolder(private val binding: TitleTodayScreenBinding) : RecyclerView.ViewHolder(binding.root) {
     fun onBind(item: TodayData.TodayTitle){
